Rename to translators -> translateUrl
